0.04a

+++ modules/customer/memo.php
@@ -10,8 +10,9 @@
 #																	#
 ####################################################
 
-$smarty->assign('customer_id', 	$VAR['customer_id']);
-$smarty->assign('customer_name', 	$VAR['customer_name']);
+$smarty->assign('customer_id', isset($VAR['customer_id']) ? $VAR['customer_id'] : '');
+$smarty->assign('customer_name', isset($VAR['customer_name']) ? $VAR['customer_name'] : '');
+
 
 if(isset($VAR['submit'])) {
 
@@ -24,10 +25,11 @@ if(isset($VAR['submit'])) {
 		force_page('core', 'error&error_msg=MySQL Error: '.$db->ErrorMsg().'&menu=1&type=database');
 		exit;
 	}
+
 	force_page('customer', 'customer_details&page_title='.$VAR['customer_name'].'&customer_id='.$VAR['customer_id']);
 } else {
 
-	if($VAR['action'] == 'delete') {
+	if(isset($VAR['action']) && $VAR['action'] == 'delete') {
 		$q = "DELETE FROM ".PRFX."CUSTOMER_NOTES WHERE ID=".$db->qstr( $VAR['note_id'] );
 
 		if(!$rs = $db->execute($q)) {
